refactor(api): extract weather point lookup into shared api.core.weather module
Move get_weather_data_for_point (and its _to_numpy_datetime64 helper) out of
api.fires.scoring into api.core.weather so both the fire-scoring pipeline and
the risk-grid module can import it without a cross-package dependency on fire
internals. Also removes a TOCTOU file-existence check and a redundant spatial
re-selection in the precipitation branch.
